Error refactoring
